requirements.txt
streamlit
PyPDF2
python-dotenv
google-generativeai
sentence-transformers
numpy==1.24.4
